fix(cc-widgets): review-comments
