+++ docs/web-component-samples/src/index.ts
@@ -1,4 +1,3 @@
-import {store, WebUserState, WebStationLogin} from '@webex/cc-widgets';
-
-//  @ts-ignore
+import {store} from '@webex/cc-widgets/wc';
+// @ts-ignore
 window.store = store;
