+++ docs/web-component-samples/webpack.config.js
@@ -4,6 +4,7 @@ const {merge} = require('webpack-merge');
 const baseConfig = require('../../webpack.config');
 
 module.exports = merge(baseConfig, {
+  entry: './src/index.ts',
   output: {
     path: path.resolve(__dirname, 'dist'),
     filename: 'bundle.js',
